+++ src/pages/attitalk/Search.jsx
@@ -23,6 +23,7 @@ const SearchContainer = styled.div`
   display: flex;
   flex-direction: column;
   width: 100%;
+  height: 100%;
   justify-content: center;
   align-items: center;
 `;
